// dispenser/src/interfaces/recipe.ts
export interface IRecipe {
  MV_RECETA_ID: string;
  TB_ENCABEZADO_TIPO_ID: string;
  TB_ENCABEZADO_TIPO_DESC: string;
  TB_ENCABEZADO_ESTADO_ID: string;
  TB_ENCABEZADO_ESTADO_DESC: string;
  TB_ENCABEZADO_ATENCION_ID: string;
  TB_ENCABEZADO_ATENCION_DESC: string;
  TB_ESTABLECIMIENTO_ID: string;
  TB_ESTABLECIMIENTO_DESC: string;
  MV_PERSONA_PACIENTE_ID: string;
  MV_PERSONA_PROFESIONAL_ID: string;
  TB_ESPECIALIDAD_ID: string;
  FECHA_ATENCION: string;
  FECHA_DIGITACION: string;
  SOLO_FECHA_DIGITACION: string;
  FECHA_VENCIMIENTO: string;
  FECHA_ENTREGA: string;
  FECHA_ENTREGA_SIN_HORA: string;
  FECHA_PRIMERA_ENTREGA: string;
  FECHA_PRIMERA_ENTREGA_SIN_HORA: string;
  PROCEDENCIA_ID: string;
  PROCEDENCIA_DESC: string;
  OBSERVACION: string;
  MV_ADMISION_URGENCIA_ID: string;
  RCE_MV_DETALLE_ID: string;
  MV_EPISODIO_ID: string;
}